mceの売り数も表示 日本語数値表記の32bitばぐ修正

+++ src/main/java/red/man10/Utility.java
@@ -30,7 +30,7 @@ public class Utility {
 
 
     public static String getJpBal(double balance){
-        int val = (int)balance;
+        long val = (long)balance;
         String addition = "";
         String form = "万";
         long man = val/10000;
